requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
python-dotenv==0.21.0
gunicorn==23.0.0
Flask-Cors==3.0.10